add a test on checkout

+++ adlplug/opl3bankeditor-source.sh
@@ -6,6 +6,10 @@
 git clone https://github.com/Wohlstand/OPL3BankEditor
 cd OPL3BankEditor
 git checkout $1
+if [ $? == 1 ]; then
+    echo "Wrong branch / tag name: $1"
+    exit 1
+fi
 git submodule update --init --recursive --progress
 find . -name .git -exec rm -rf {} \;
 cd ..
