+++ chmaba/rcverb-source.sh
@@ -5,6 +5,10 @@
 git clone https://github.com/chmaha/RCVerb
 cd RCVerb
 git checkout $1
+if [ $? == 1 ]; then
+    echo "Wrong branch / tag name: $1"
+    exit 1
+fi
 git submodule update --init --recursive --progress
 find . -name .git -exec rm -rf {} \;
 cd ..
