+++ string-machine/string-machine-source.sh
@@ -6,6 +6,10 @@
 git clone https://github.com/jpcima/string-machine
 cd string-machine
 git checkout $1
+if [ $? == 1 ]; then
+    echo "Wrong branch / tag name: $1"
+    exit 1
+fi
 git submodule update --init --recursive --progress
 find . -name .git -exec rm -rf {} \;
 cd ..
