+++ zl/zl-source.sh
@@ -6,6 +6,10 @@
 git clone https://github.com/ZL-Audio/$1
 cd $1
 git checkout $2
+if [ $? == 1 ]; then
+    echo "Wrong branch / tag name: $2"
+    exit 1
+fi
 git submodule update --init --recursive --progress
 find . -name .git -exec rm -rf {} \;
 cd ..
